Fix `same subnet in different sections` bug (#33)
The lookup for subnets now respect the provided section.
Now a subnet can be used in different sections.

To test more complex situations we refactored the test data and
test scripts to be more common for all test cases.

The condition in the `run` method for deleting children has not workded
properly. We fix these condition to use the correct method to resolve
parent.

Add changelog fragment.