Prepare splitting phpipam stuff into own class
As we need interaction with phpipam api not only in modules but also in
plugins like lookup and inventory plugins we decided to cut all the
common stuff out into a seperate class.
This class should provide all methods to interact with the phpipam api
in one place.
With this first commit we start to adapt all existing methods from
`phpipam_helper` module in `_phpypam` module. Next we put comments into
the source module to made clear what to achieve with the new class and
its methods.
This commit is non inversive.