Fix vlan to id resolve mechanism
As reported in #80 it is currently not possible to resolve a vlan number
to its id. This relies on a different structure of vlan entities. Vlan
entities has `vlanId` instead of `id` as usual.
We decided to use the same approach as for device_type to add `vlanId`
as `id` to the resolved entity.