Add subnet parent test
As there is some work needed in managing subnets and/or folders and its
parents a test case is needed.
For that purpose we added a test playbook and a needed variable for
example data.
We also fix a bug in the `subnet` task to.
With all these changes/additions we can test subenets and parent subnets
first. In later steps we need to extend the test to also check the
majority of all combinations of subnets, folders and parents.